moving average

Create a new map/matrix by doing a moving average on a matrix

moving_average(Z, r)

Arguments

Z
matrix/dataframe

r
integer : the radius of the window of the moving average. Ex : r = 1 implies a window of 3*3

Value

matrix (side effects imply a smaller matrix than the Z matrix passed in parametre)

Examples

moving_average(matrix(rnorm(10*10, 0,1), nrow = 10), r=2)

#>           [,1]        [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,] 0.3692207  0.15951269 0.10660446  0.21496700 -0.04294576 -0.05699105
#> [2,] 0.3719038  0.16255316 0.28508211  0.23297583  0.07163706  0.06702227
#> [3,] 0.1596508  0.04061091 0.16690480  0.10806089  0.03509572  0.07548343
#> [4,] 0.1377411 -0.07604166 0.01183075 -0.05530150 -0.10683186 -0.11682580
#> [5,] 0.1986293  0.04096930 0.12710875  0.02436386  0.06573003 -0.08948269
#> [6,] 0.1041015  0.04847889 0.17566336  0.13264592  0.20784849  0.10605873